+++ lib/instream/query/write.ex
@@ -5,8 +5,8 @@ defmodule Instream.Query.Write do
 
   use Instream.Query
 
-  def execute(%Query{ payload: payload }, opts, conn) do
-    payload
+  def execute(%Query{} = query, opts, conn) do
+    query
     |> conn[:writer].write(opts, conn)
     |> maybe_parse(opts)
   end